rag-engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sentence-transformers==3.3.1
chromadb==0.6.3
pypdf2==3.0.1
python-docx==1.1.2
pydantic==2.10.4
numpy==1.26.4
torch==2.5.1
transformers==4.47.1
openai==1.12.0
